fix #1643 and improve integration test (#1645)
